+++ src/AbstractExecutable.ts
@@ -8,7 +8,7 @@ import { AbstractTest } from './AbstractTest';
 import { TaskPool } from './util/TaskPool';
 import { WorkspaceShared } from './WorkspaceShared';
 import { ExecutableRunResultValue, RunningExecutable } from './RunningExecutable';
-import { promisify, inspect } from 'util';
+import { promisify } from 'util';
 import { Version, getAbsolutePath, CancellationToken, CancellationFlag, reindentStr } from './Util';
 import {
   resolveOSEnvironmentVariables,
@@ -23,6 +23,7 @@ import { debugAssert, debugBreak } from './util/DevelopmentHelper';
 import { SpawnBuilder } from './Spawner';
 import { SharedTestTags } from './SharedTestTags';
 import { Disposable } from './Util';
+import { FilePathResolver, TestItemParent } from './TestItemManager';
 
 export class TestsToRun {
   public readonly direct: AbstractTest[] = []; // test is drectly included, should be run even if it is skipped
@@ -34,7 +35,7 @@ export class TestsToRun {
   }
 }
 
-export abstract class AbstractExecutable implements Disposable {
+export abstract class AbstractExecutable implements Disposable, FilePathResolver {
   public constructor(
     public readonly shared: WorkspaceShared,
     public readonly properties: RunnableProperties,
@@ -68,8 +69,6 @@ export abstract class AbstractExecutable implements Disposable {
 
   private _lastReloadTime: number | undefined = undefined;
 
-  //TODO:future  special group to be expandable: private _execGroup: vscode.TestItem | undefined = undefined;
-
   // don't use this directly because _addTest and _getTest can be overwritten
   private _tests = new Map<string /*id*/, AbstractTest>();
 
@@ -81,20 +80,26 @@ export abstract class AbstractExecutable implements Disposable {
     return this._tests.get(testId) as T;
   }
 
-  private _getOrCreateChildGroup(
+  private async _getOrCreateChildGroup(
     idIn: string | undefined,
     label: string,
     description: string,
-    _tooltip: string, // tooltip currently is not supported
     itemOfLevel: vscode.TestItem | undefined,
-  ): vscode.TestItem {
-    const childrenOfLevel = itemOfLevel?.children ?? this.shared.rootItems;
+  ): Promise<vscode.TestItem> {
+    const childrenOfLevel = this.shared.testController.getChildCollection(itemOfLevel);
     const id = idIn ?? label;
     const found = childrenOfLevel.get(id);
     if (found) {
       return found;
     } else {
-      const testItem = this.shared.testItemCreator(id, label, undefined, undefined, undefined);
+      const testItem = await this.shared.testController.createOrReplace(
+        itemOfLevel,
+        id,
+        label,
+        undefined,
+        undefined,
+        undefined,
+      );
       testItem.description = description;
       testItem.tags = SharedTestTags.groupArray;
       childrenOfLevel.add(testItem);
@@ -107,14 +112,12 @@ export abstract class AbstractExecutable implements Disposable {
     id: string | undefined,
     label: string,
     description: string | undefined,
-    tooltip: string | undefined,
     varsToResolve: ResolveRuleAsync<string>[],
   ): Promise<vscode.TestItem> {
     const resolvedLabel = await this.resolveText(label, ...varsToResolve);
     const resolvedDescr = description !== undefined ? await this.resolveText(description, ...varsToResolve) : '';
-    const resolvedToolt = tooltip !== undefined ? await this.resolveText(tooltip, ...varsToResolve) : '';
 
-    return this._getOrCreateChildGroup(id, resolvedLabel, resolvedDescr, resolvedToolt, itemOfLevel);
+    return this._getOrCreateChildGroup(id, resolvedLabel, resolvedDescr, itemOfLevel);
   }
 
   private _updateVarsWithTags(tg: TestGrouping, tags: string[], tagsResolveRule: ResolveRuleAsync<string>): void {
@@ -167,14 +170,13 @@ export abstract class AbstractExecutable implements Disposable {
   protected async _createTreeAndAddTest<T extends AbstractTest>(
     testGrouping: TestGrouping,
     testId: string,
-    file: string | undefined,
+    resolvedFile: string | undefined,
     tags: string[], // in case of google test it is the TestCase
     _description: string | undefined, // currently we don't use it for subtree creation
-    createTest: (container: vscode.TestItemCollection) => T,
+    createTest: (parent: TestItemParent) => T,
     updateTest: (test: T) => void,
   ): Promise<T> {
-    this.shared.log.info('testGrouping', testId);
-    this.shared.log.debug('testGrouping', { testId, file, tags, testGrouping });
+    this.shared.log.info('testGrouping', { testId, resolvedFile, tags, testGrouping });
 
     tags.sort();
 
@@ -182,12 +184,12 @@ export abstract class AbstractExecutable implements Disposable {
       resolve: AbstractExecutable._tagVar,
       rule: '', // will be filled soon enough
     };
-    const sourceRelPath = file ? pathlib.relative(this.shared.workspaceFolder.uri.fsPath, file) : '';
+    const sourceRelPath = resolvedFile ? pathlib.relative(this.shared.workspaceFolder.uri.fsPath, resolvedFile) : '';
 
     const varsToResolve = [
       tagsResolveRule,
       createPythonIndexerForPathVariable('sourceRelPath', sourceRelPath),
-      createPythonIndexerForPathVariable('sourceAbsPath', file ? file : ''),
+      createPythonIndexerForPathVariable('sourceAbsPath', resolvedFile ? resolvedFile : ''),
     ];
 
     // undefined means root
@@ -204,23 +206,17 @@ export abstract class AbstractExecutable implements Disposable {
           const label = g.label !== undefined ? g.label : '${filename}';
           const description = g.description !== undefined ? g.description : '${relDirpath}${osPathSep}';
 
-          itemOfLevel = await this._resolveAndGetOrCreateChildGroup(
-            itemOfLevel,
-            id,
-            label,
-            description,
-            `Path: ${this.properties.path}\nCwd: ${this.properties.options.cwd}`,
-            varsToResolve,
-          );
+          itemOfLevel = await this._resolveAndGetOrCreateChildGroup(itemOfLevel, id, label, description, varsToResolve);
 
-          this._execItem.item = itemOfLevel;
+          // special item handling for exec
+          this._execItem.setItem(itemOfLevel, resolvedFile);
 
           currentGrouping = g;
         } else if (currentGrouping.groupBySource) {
           const g = currentGrouping.groupBySource;
           this._updateVarsWithTags(g, tags, tagsResolveRule);
 
-          if (file) {
+          if (resolvedFile) {
             const label = g.label ? g.label : sourceRelPath;
             const description = g.description;
 
@@ -229,15 +225,26 @@ export abstract class AbstractExecutable implements Disposable {
               undefined,
               label,
               description,
-              undefined,
               varsToResolve,
             );
+
+            // special item handling for source file
+            if (resolvedFile && itemOfLevel.uri === undefined) {
+              itemOfLevel = await this.shared.testController.update(
+                itemOfLevel,
+                resolvedFile,
+                undefined,
+                this,
+                null,
+                null,
+                null,
+              );
+            }
           } else if (g.groupUngroupedTo) {
             itemOfLevel = await this._resolveAndGetOrCreateChildGroup(
               itemOfLevel,
               undefined,
               g.groupUngroupedTo,
-              undefined,
               undefined,
               varsToResolve,
             );
@@ -260,7 +267,6 @@ export abstract class AbstractExecutable implements Disposable {
                   undefined,
                   g.label ? g.label : AbstractExecutable._tagVar,
                   g.description,
-                  undefined,
                   varsToResolve,
                 );
               } else if (g.groupUngroupedTo) {
@@ -268,7 +274,6 @@ export abstract class AbstractExecutable implements Disposable {
                   itemOfLevel,
                   undefined,
                   g.groupUngroupedTo,
-                  undefined,
                   undefined,
                   varsToResolve,
                 );
@@ -284,7 +289,6 @@ export abstract class AbstractExecutable implements Disposable {
                   undefined,
                   g.label ? g.label : AbstractExecutable._tagVar,
                   g.description,
-                  undefined,
                   varsToResolve,
                 );
               } else if (g.groupUngroupedTo) {
@@ -292,7 +296,6 @@ export abstract class AbstractExecutable implements Disposable {
                   itemOfLevel,
                   undefined,
                   g.groupUngroupedTo,
-                  undefined,
                   undefined,
                   varsToResolve,
                 );
@@ -351,7 +354,6 @@ export abstract class AbstractExecutable implements Disposable {
                   undefined,
                   label,
                   description,
-                  undefined,
                   varsToResolve,
                 );
               } else if (g.groupUngroupedTo) {
@@ -359,7 +361,6 @@ export abstract class AbstractExecutable implements Disposable {
                   itemOfLevel,
                   undefined,
                   g.groupUngroupedTo,
-                  undefined,
                   undefined,
                   varsToResolve,
                 );
@@ -379,17 +380,16 @@ export abstract class AbstractExecutable implements Disposable {
       this.shared.log.exceptionS(e);
     }
 
-    const childrenOfLevel = itemOfLevel?.children ?? this.shared.rootItems;
-    const found = childrenOfLevel.get(testId);
+    const found = this.shared.testController.getChildCollection(itemOfLevel).get(testId);
 
     if (found) {
-      const test = this.shared.testItemMapper(found) as T;
+      const test = this.shared.testController.map(found) as T;
       if (!test) throw Error('missing test for item');
       updateTest(test);
       this._addTest(test.id, test);
       return test;
     } else {
-      const test = createTest(childrenOfLevel);
+      const test = createTest(itemOfLevel);
       this._addTest(test.id, test);
       return test;
     }
@@ -398,12 +398,11 @@ export abstract class AbstractExecutable implements Disposable {
   private removeWithLeafAscendants(testItem: vscode.TestItem, evenIfHasChildren = false): void {
     if (!evenIfHasChildren && testItem.children.size > 0) return;
 
-    if (testItem.parent) {
-      const parent = testItem.parent;
-      parent.children.delete(testItem.id);
+    const parent = testItem.parent;
+    this.shared.testController.getChildCollection(parent).delete(testItem.id);
+
+    if (parent) {
       this.removeWithLeafAscendants(parent);
-    } else {
-      this.shared.rootItems.delete(testItem.id);
     }
   }
 
@@ -411,12 +410,12 @@ export abstract class AbstractExecutable implements Disposable {
     this.removeWithLeafAscendants(test.item, true);
   }
 
-  protected _createAndAddError(label: string, message: string): void {
-    this._execItem.setError(label, message);
+  protected async _createAndAddError(label: string, message: string): Promise<void> {
+    await this._execItem.setError(label, message);
   }
 
-  protected _createAndAddUnexpectedStdError(stdout: string, stderr: string): void {
-    this._createAndAddError(
+  protected async _createAndAddUnexpectedStdError(stdout: string, stderr: string): Promise<void> {
+    await this._createAndAddError(
       `⚡️ Unexpected ERROR while parsing`,
       [
         `❗️Unexpected stderr!`,
@@ -544,15 +543,6 @@ export abstract class AbstractExecutable implements Disposable {
     taskPool: TaskPool,
     cancellationToken: CancellationToken,
   ): Promise<void> {
-    // DISABLED for now
-    // try {
-    //   await this.runTasks('beforeEach', taskPool, cancellationToken);
-    //   //await this.reloadTests(taskPool, cancellationToken); // this might relod the test list if the file timestamp has changed
-    // } catch (e) {
-    //   //this.sentStaticErrorEvent(testRunId, collectChildrenToRun(), e);
-    //   return;
-    // }
-
     const testsToRunFinal: AbstractTest[] = [];
 
     for (const t of testsToRun.direct) {
@@ -566,6 +556,20 @@ export abstract class AbstractExecutable implements Disposable {
 
     if (testsToRunFinal.length == 0) return;
 
+    try {
+      await this.runTasks('beforeEach', taskPool, cancellationToken);
+      //TODO: future: test list might changes: await this.reloadTests(taskPool, cancellationToken);
+      // that case the testsToRunFinal should be after this block
+    } catch (e) {
+      const msg = e.toString();
+      testRun.appendOutput(msg);
+      const errorMsg = new vscode.TestMessage(msg);
+      for (const test of testsToRun) {
+        testRun.errored(test.item, errorMsg);
+      }
+      return;
+    }
+
     const buckets = this._splitTestSetForMultirunIfEnabled(testsToRunFinal);
     await Promise.allSettled(
       buckets.map(async (bucket: readonly AbstractTest[]) => {
@@ -576,7 +580,13 @@ export abstract class AbstractExecutable implements Disposable {
     try {
       await this.runTasks('afterEach', taskPool, cancellationToken);
     } catch (e) {
-      //this.sentStaticErrorEvent(testRunId, collectChildrenToRun(), e);
+      const msg = e.toString();
+      testRun.appendOutput(msg);
+      const errorMsg = new vscode.TestMessage(msg);
+      for (const test of testsToRun) {
+        testRun.errored(test.item, errorMsg);
+      }
+      return;
     }
   }
 
@@ -759,9 +769,7 @@ export abstract class AbstractExecutable implements Disposable {
 
             if (exitCode !== undefined) {
               if (exitCode !== 0) {
-                throw Error(
-                  `Task "${taskName}" has returned with exitCode(${exitCode}) != 0. (\`testMate.test.advancedExecutables:runTask.${type}\`)`,
-                );
+                throw Error(`Task "${taskName}" has returned with exitCode(${exitCode}) != 0.`);
               }
             }
           }
@@ -812,34 +820,6 @@ export abstract class AbstractExecutable implements Disposable {
     return found || path;
   }
 
-  public sendStaticEvents(
-    _testRunId: string,
-    _childrenToRun: readonly AbstractTest[],
-    _staticEvent: unknown | undefined,
-  ): void {
-    // childrenToRun.forEach(test => {
-    //   const testStaticEvent = test.getStaticEvent(testRunId);
-    //   const event: TestEvent | undefined = staticEvent || testStaticEvent;
-    //   if (event) {
-    //     event.test = test;
-    //     event.testRunId = testRunId;
-    //     // we dont need to send events about ancestors: https://github.com/hbenl/vscode-test-explorer/issues/141
-    //     // probably we dont need this either: this._shared.sendTestEvent(test!.getStartEvent());
-    //     this._shared.sendTestRunEvent(event);
-    //   }
-    // });
-  }
-
-  // eslint-disable-next-line
-  public sentStaticErrorEvent(testRunId: string, childrenToRun: readonly AbstractTest[], err: any): void {
-    this.sendStaticEvents(testRunId, childrenToRun, {
-      type: 'test',
-      test: 'will be filled automatically',
-      state: 'errored',
-      message: err instanceof Error ? `⚡️ ${err.name}: ${err.message}` : inspect(err),
-    });
-  }
-
   protected processStdErr(testRun: vscode.TestRun, runPrefix: string, str: string): void {
     testRun.appendOutput(runPrefix + '⬇ std::cerr:\r\n');
     const indented = reindentStr(0, 2, str);
@@ -857,28 +837,32 @@ export interface HandleProcessResult {
 class ExecutableGroup {
   public constructor(private readonly executable: AbstractExecutable) {}
 
-  private _count = 0;
+  private _busyCounter = 0;
   private _item: vscode.TestItem | undefined = undefined;
   private _itemForStaticError: vscode.TestItem | undefined = undefined;
   // we need to be exclusive because we save prevTests
   private _lock = Promise.resolve();
 
-  public set item(item: vscode.TestItem) {
-    if (this.item && this.item !== item) {
+  public setItem(item: vscode.TestItem, _resolvedFile: string | undefined) {
+    if (this._item && this._item !== item) {
       this.executable.shared.log.errorS('why do we have different executableItem');
       debugBreak('why are we here?');
     } else if (!this._item) {
       this._item = item;
-      if (this._count > 0) this._item.busy = true;
+      if (this._busyCounter > 0) this._item.busy = true;
     }
+
+    // if (resolvedFile) {
+    //   //TODO if (this._item)
+    // }
   }
 
   // makes the item spinning
   public busy(func: () => Promise<void>): Promise<void> {
-    if (this._count++ === 0 && this._item) this._item.busy = true;
+    if (this._busyCounter++ === 0 && this._item) this._item.busy = true;
 
     return (this._lock = this._lock.then(func).finally(() => {
-      if (--this._count === 0) {
+      if (--this._busyCounter === 0) {
         if (this._item) {
           this._item.busy = false;
         }
@@ -886,21 +870,21 @@ class ExecutableGroup {
     }));
   }
 
-  public setError(label: string, message: string): void {
+  public async setError(label: string, message: string): Promise<void> {
     const l = label + ': ' + message;
     if (this._item) {
       this._item.error = l;
       this.removeSpecialItem();
     } else {
       if (!this._itemForStaticError) {
-        this._itemForStaticError = this.executable.shared.testItemCreator(
+        this._itemForStaticError = await this.executable.shared.testController.createOrReplace(
+          undefined,
           this.executable.properties.path,
           this.executable.properties.path,
           undefined,
           undefined,
           undefined,
         );
-        this.executable.shared.rootItems.add(this._itemForStaticError);
       }
       this._itemForStaticError.error = l;
     }
@@ -913,7 +897,9 @@ class ExecutableGroup {
 
   private removeSpecialItem(): void {
     if (this._itemForStaticError) {
-      this.executable.shared.rootItems.delete(this._itemForStaticError.id);
+      this.executable.shared.testController
+        .getChildCollection(this._itemForStaticError.parent)
+        .delete(this._itemForStaticError.id);
       this._itemForStaticError = undefined;
     }
   }