+++ CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.0.9] - 2021-12-02
+
+### Fixed
+
+- an issue that `testMate.cpp.test.advancedExecutables` framework specific settigns were only picking `catch2`
+
+## [4.0.8] - 2021-12-01
+
+### Added
+
+- `envFile` now supports `.env` format too. (determines from the end of the filename)
+- removing missing SubCases / Sections
+
+## [4.0.7] - 2021-11-27
+
+### Fixed
+
+- running tasks feature with limitation: test list of executables won't be reloaded so in case the task results in a new test then that test won't be run.
+
 ## [4.0.6] - 2021-11-25
 
 - Improved Output formatting and colorization